requirements.txt
langgraph>=1.0.0
langchain>=1.0.0
langchain-google-genai>=2.1.0
langchain-community>=0.3.0
tavily-python>=0.5.0
langgraph-checkpoint-sqlite>=1.0.0
streamlit>=1.40.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
